bots/finance_bot/requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
python-fints>=3.1.0
requests>=2.31.0
cryptography>=41.0.0 